Fix a security issue.
streamin~ was vulnerable to a remotely exploitable heap buffer overflow that
could potentially allow an attacker to execute arbitrary code on a machine
that is using streamin~.  The issue is that the defined frame header includes
a field for frame size which indicates how much additional data is available.
However, this frame size was not checked to ensure that it was smaller than
the size of the buffer that was allocated to store this data.

(one of my patches from issue #1848356)


git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@9133 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d