(add patch from issue #1852385)
This patch provides a revised implementation of the strtokcpy() function in
s_path.c. It provides the following benefits:

1) Prevent potential overflow of a stack buffer. This function did nothing
to ensure that it didn't write past the end of the destination buffer.

It is possible to cause this to happen by providing certain command line
arguments that are longer than MAXPDSTRING. Also, there may be other ways
to trigger this bug if namelist_append_files() is used anywhere beyond the
uses I reviewed, which are the ones in pd/*.c.

2) Copy bytes from the string in the same loop that looks for the
delimiter. This is simply for efficiency in that the string only has to be
traversed once, instead of twice (one to find the delimiter, and the second
to copy up to it).


git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@9143 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d