This commit adds configure script checking for a number of functions in libm: - sinf, cosf, atanf, atan2f, ...
Previously, in x_arithmetic.c, these functions were redefined for MSW /
Apple. Now, they will only be redefined if the configure script does not
find them.

The results of the checks for these functions are placed in the config.h
header file. The config.h.in file was generated using autoheader.

(Note that this patch requires the inclusion of acinclude.m4. You must run
 aclocal before running autoconf to generate the configure script.)


git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@9144 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d