patch #1963983
On Windows, FILENAME_MAX is much smaller than MAXPDSTRING, so I replaced MAXPDSTRING with FILENAME_MAX everywhere I could find that is related to filenames.  FILENAME_MAX is a POSIX standard macro for defining the max length of a complete filename.

The current situation could result in crashes on Windows.


git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@9797 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d